refactored balance line sort into generics
