+++ src/components/AuditTrail/AuditTrailReducer.js
@@ -1,7 +1,7 @@
 import { AUDITTRAIL_STREAM, GET_AUDITTRAIL } from "./AuditTrailActions";
 import { SUCCESS_POSTFIX } from "../../api/QueryActions";
 import { STOP_STREAM } from "api/StreamActions";
-import { guidToObjectKey, copyProperties, dlog } from "utils/Generics";
+import { guidToObjectKey, copyProperties, balanceLineSort } from "utils/Generics";
 
 
 // State definition
@@ -10,8 +10,10 @@ const initialState = {
   auditTrails: {
     // key = initiativeId|actorId|GENERIC, value = audittrail
   },
-  integralAuditTrail: null,   // One audittrail constructed from all audit trails with the records sorted on date (past to future) and the totalRecords being the sum of all totalRecords
-                              // Can be set to null at any time
+  integralAuditTrail: {   // One audittrail constructed from all audit trails with the records sorted on date (past to future) and the totalRecords being the sum of all totalRecords
+    records: [],
+    totalRecords: 0
+  }
 }
 
 // The audittrail object has this structure:
@@ -88,37 +90,15 @@ const processRecords = records => records.map(record => {
 
 const constructIntegralAuditTrail = (auditTrails) => {
   const recordsList = []
-  let indexes = []
   let grantTotalRecords = 0
+  
   for (let i in auditTrails) {
     recordsList.push(auditTrails[i].records)
     grantTotalRecords += auditTrails[i].totalRecords
-    indexes.push(0)
   }
 
-  let sortedRecords = []
+  const sortedRecords = balanceLineSort(recordsList, (record1, record2) => record1.instant < record2.instant)
 
-  let nextRecordListIdx = 0
-  var compareDate = null
-  while (nextRecordListIdx > -1) {
-    nextRecordListIdx = -1
-    compareDate = null
-    for (let i = 0; i < indexes.length; i++) {
-      const idx = indexes[i]
-      const records = recordsList[i]
-      if (idx < records.length) {
-        if (nextRecordListIdx === -1 || records[idx].instant < compareDate) {
-          nextRecordListIdx = i
-          compareDate = records[idx].instant
-        }
-      }
-    }
-
-    if (nextRecordListIdx > -1) {
-      sortedRecords.push(recordsList[nextRecordListIdx][indexes[nextRecordListIdx]])
-      indexes[nextRecordListIdx] = indexes[nextRecordListIdx] + 1
-    }
-  }
   return {
     records: sortedRecords,
     totalRecords: grantTotalRecords
